Move hasAKI
